requirements.txt
absl-py==0.12.0
asgiref==3.2.10
astor==0.8.1
astroid==1.5.3
astunparse==1.6.3
attrs==19.3.0
backports.functools-lru-cache==1.4
beautifulsoup4==4.6.0
boto3==1.14.31
botocore==1.17.31
bs4==0.0.1
cachetools==4.0.0
certifi==2017.4.17
chardet==3.0.4
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.4
configparser==3.5.0
cv2-tools==2.4.0
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
decorator==4.0.11
dill==0.3.1.1
Django==3.0.8
django-appconf==1.0.4
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-htmlmin==0.11.0
django-js-asset==1.2.2
django-sass-processor==0.8.2
django-storages==1.9.1
django-user-agents==0.4.0
djangorestframework==3.12.2
docupy==0.3.2
docutils==0.15.2
easyocr==1.1.3
enum34==1.1.6
et-xmlfile==1.0.1
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.1
idna==2.5
ImageHash==4.0
imageio==2.9.0
imbalanced-learn==0.6.2
imblearn==0.0
imutils==0.5.3
isort==4.2.15
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.14.1
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
libsass==0.20.1
Markdown==3.2
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mediapipe==0.8.0
mock==4.0.3
msgpack-python==0.5.6
networkx==1.11
nltk==3.5
numpy==1.19.3
oauthlib==3.1.0
olefile==0.44
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
openpyxl==3.0.5
opt-einsum==3.3.0
pandas==1.0.3
pathlib2==2.3.5
Pillow==8.1.0
Pillow-PIL==0.1.dev0
pprint==0.1
promise==2.3
protobuf==3.14.0
psutil==5.8.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.0.0.dev6
pylinktester==0.1
pylint==1.7.2
pymongo==3.11.2
pyparsing==2.2.0
pyperclip==1.8.0
python-bidi==0.4.2
python-constraint==1.4.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
pywin32==227
pyWinhook==1.6.2
PyYAML==5.3
redis==3.5.3
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
reverse-geocoder==1.5.1
rjsmin==1.1.0
rsa==4.0
s3transfer==0.3.3
scikit-image==0.18.1
scikit-learn==0.22.1
scipy==1.1.0
seaborn==0.10.0
singledispatch==3.4.0.3
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
sqlparse==0.3.1
tblib==1.7.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboard-pytorch==0.7.1
tensorflow==2.4.0
tensorflow-datasets==2.1.0
tensorflow-estimator==2.4.0
tensorflow-metadata==0.21.1
termcolor==1.1.0
tifffile==2020.12.8
torch==1.2.0+cpu
torchvision==0.4.0+cpu
tqdm==4.44.0
typing-extensions==3.7.4.3
ua-parser==0.10.0
urllib3==1.21.1
user-agents==2.2.0
utils==1.0.1
webencodings==0.5.1
Werkzeug==1.0.0
whitenoise==5.1.0
wordcloud==1.6.0
wrapt==1.12.1
wxPython==4.1.1
